requirements.txt
# Run these libraries before accessing model
flask
flask-cors
groq
SpeechRecognition
transformers
textblob
termcolor
python-dotenv
torch
torchvision
torchaudio
tiktoken
protobuf
happytransformer
gradio>=4.0.0
spacy
nltk
librosa
numpy
pandas
matplotlib
seaborn